Renamed the image class to Poster and added Pillow to process images. Poster class can get Images now
