+++ requirements.txt
@@ -16,6 +16,7 @@ MarkupSafe==0.23
 path.py==8.1.2
 pexpect==4.0.1
 pickleshare==0.5
+Pillow==3.0.0
 ptyprocess==0.5
 Pygments==2.0.2
 pytz==2015.7
