+++ tmdbcall/__init__.py
@@ -1,4 +1,4 @@
 """
 This module gets data from TheMovieDataBase
 """
-from . import tv, job, genre, person, image
+from . import tv, job, genre, person, poster
